requirements.txt
arrow==1.2.3
asgiref==3.7.2
Babel==2.12.1
blessed==1.20.0
Django==4.2.2
django-cors-headers==4.1.0
django-phonenumber-field==7.1.0
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.14.0
phonenumbers==8.13.14
python-dateutil==2.8.2
pytz==2023.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
wcwidth==0.2.6
